pyproject.toml
[tool.poetry]
name = "er"
version = "0.1.0"
description = "Early Retirement simulation"
authors = ["Andrea Dal Pozollo <dalpozz@gmail.com>"]
license = "Apache 2.0"
repository = "https://github.com/dalpozz/early-retirement"
keywords = ["streamlit", "fire", "early", "retirement"]
readme = "README.md"


[tool.poetry.dependencies]
python = ">3.11,<3.12"
streamlit = "^1.41.1"
pandas = "^2.2.3"
pyyaml = "^6.0.2"
numpy = "^2.2.2"
numpy-financial = "^1.0.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
# https://github.com/psf/black
line-length = 100

[tool.isort]
# https://github.com/timothycrosley/isort/
known_typing = "typing,types,typing_extensions,mypy,mypy_extensions"
sections = "FUTURE,TYPING,STDLIB,THIRDPARTY,FIRSTPARTY,LOCALFOLDER"
include_trailing_comma = true
default_section = "FIRSTPARTY"
multi_line_output = 3
indent = 4
force_grid_wrap = 0
use_parentheses = true
line_length = 100
